removed remapping pine by default as it causes issues in sxmo
It seems that sxmo binds LWin to Super_L and binding it to something
else here overrides that and breaks Sway shortcuts, and binding the
Sway $mod key to anything other than Mod# (like Meta_L) didn't work,
so I'll leave it unchanged in the default layout and only override it
in layouts that require doing so. This also means that I moved the
default Pine layout to a separate one to still be able to include a
default layout that doesn't redefine Pine. So the default/basic layout
is now basically just a template for and to be included in other layouts and not to actually be used.