fix: tty access should now be as consistent and working as possible
There are two issues: 1. X/Wayland always takes the hardware F# keys
for the TTY switch bindings and not what is redefined to be F# in xkb,
and 2. depending on the modifier keys used, the electrical matrix of
the keyboard makes some combinations inaccessible. Therefore I bound
F# hardware keys in the userspace layouts to the number keys to always
have TTY switch bindings work with those keys, and in TTY keymaps that
do not use hardware F# keycodes for F# software keycodes but a regular
modifier like Pine or AlrGr, I additionally mapped TTY switch binds to
hardware F# keys as well to make it consistent with xkb where the
rebound F# keys can't be used to switch TTY.